+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack serve --port 3000  ",
-    "build": "gulp build",
+    "build": "gulp build && npm run replace",
     "replace": "replace.bat"
   },
   "author": "Ramakrishnan",
@@ -29,28 +29,29 @@
   "devDependencies": {
     "@types/webpack": "5.28.0",
     "babel-loader": "^8.2.2",
-    "css-loader": "6.2.0",
-    "html-webpack-plugin": "5.3.2",
-    "mini-css-extract-plugin": "2.1.0",
-    "ts-node": "10.1.0",
-    "typescript": "4.3.5",
-    "webpack": "5.45.1",
-    "webpack-cli": "4.7.2",
-    "webpack-dev-server": "3.11.2",
-    "html-loader": "2.1.2",
-    "less": "^3.12.2",
-    "less-loader": "10.0.1",
-    "sass-loader": "12.1.0",
-    "style-loader": "3.1.0",
-    "ts-loader": "9.2.3",
-    "url-loader": "4.1.1",
+    "css-loader": "^6.2.0",
+    "del": "6.0.0",
     "gulp": "4.0.2",
     "gulp-concat": "2.6.1",
-    "gulp-sass": "4.1.1",
     "gulp-less": "4.0.1",
-    "gulp-typescript": "5.0.1",
-    "del": "6.0.0",
-    "source-map-loader": "3.0.0"
+    "gulp-typescript": "^5.0.1",
+    "html-loader": "2.1.2",
+    "html-webpack-plugin": "5.3.2",
+    "i": "^0.3.6",
+    "less": "^3.12.2",
+    "less-loader": "10.0.1",
+    "mini-css-extract-plugin": "2.1.0",
+    "npm": "^7.20.6",
+    "sass-loader": "12.1.0",
+    "source-map-loader": "3.0.0",
+    "style-loader": "^3.1.0",
+    "ts-loader": "9.2.3",
+    "ts-node": "10.1.0",
+    "typescript": "4.3.5",
+    "url-loader": "4.1.1",
+    "webpack": "5.45.1",
+    "webpack-cli": "4.7.2",
+    "webpack-dev-server": "3.11.2"
   },
   "peerDependencies": {
     "react": "^16.14.0",
