+++ src/DraftEditor.tsx
@@ -4,15 +4,15 @@ import { convertToRaw, DraftInlineStyle, DraftStyleMap, EditorState, Modifier, R
 import 'draft-js/dist/Draft.css';
 import draftToHtml from 'draftjs-to-html';
 import React, { Component, Fragment } from 'react';
+import { SuggestionList } from './Mention'
 import DraftToolbar, { IDraftElementFormats } from './DraftToolbar/DraftToolbar';
 import './Styles';
 import createMentionPlugin, {
     defaultSuggestionsFilter,
 } from '@draft-js-plugins/mention';
-import { SuggestionList } from "./Mention";
-import { MentionDataValue } from './MentionData';
-
-
+import "@draft-js-plugins/mention/lib/plugin.css";
+import  { convertFromHTMLString, resolveCustomStyleMap, formatText, getFormat, getContentFromEditorState } from './Service/draftEditor'
+import { mentionsStyles } from './Styles';
 // import { Editor, createPlugin, pluginUtils } from "draft-extend";
 import { convertToHTML, convertFromHTML } from "draft-convert";
 
@@ -21,15 +21,15 @@ interface IDraftEditorProps {
     initialContent?: string;
     textAlignment?: string;
     onContentChange?: (content: string) => void;
-    onContentTextChange?: (content: { formattedText: string, value: string, mentionList :string[]}) => void;
+    onContentTextChange?: (content: { formattedText: string, value: string, mentionList: string[] }) => void;
     onCurrentFormatChange?: (formats: IDraftElementFormats) => void;
     showToolbar?: boolean;
     innerRef?: any;
-    showMention?: { value: boolean, people: boolean};
-    onMentionInput?:(value:string) => void;
+    showMention?: { value: boolean, people: boolean };
+    onMentionInput?: (value: string) => void;
     peopleSuggestion?: any[];
     valueSuggestion?: any[];
-    isMentionLoading?:boolean;
+    isMentionLoading?: boolean;
 }
 
 export interface IDraftEditorRef {
@@ -49,198 +49,48 @@ const CUSTOM_STYLE_MAP: DraftStyleMap = {
 };
 
 const MENTION_SUGGESTION_NAME = {
-    PREFIX_ONE:"@",
-    PREFIX_TWO:"#"
+    PREFIX_ONE: "@",
+    PREFIX_TWO: "#"
 }
 
-const resolveCustomStyleMap = (style: DraftInlineStyle) => {
-    const colObj = {} as React.CSSProperties;
-    style.forEach((styleKey) => {
-        if (styleKey) {
-            if (styleKey.includes('color-')) {
-                const [, color] = styleKey.split('color-');
-                colObj.color = color;
-            } else if (styleKey.includes('backgroundColor-')) {
-                const [, backgroundColor] = styleKey.split('backgroundColor-');
-                colObj.backgroundColor = backgroundColor;
-            } else if (styleKey.includes('fontfamily-')) {
-                const [, font] = styleKey.split('fontfamily-');
-                colObj.fontFamily = font;
-            } else if (styleKey.includes('fontsize-')) {
-                const [, size] = styleKey.split('fontsize-');
-                colObj.fontSize = size;
-            }
-        }
-    });
-    return colObj;
-};
-
-const getEditorStateFromContent = (content: string) => {
-    const importOptions = {
-        customInlineFn: (htmlElement: HTMLElement, { Style, Entity }: InlineCreators) => {
-            const styles = [];
-            if (htmlElement.style.color) {
-                styles.push(`color-${htmlElement.style.color}`);
-            }
-            if (htmlElement.style.backgroundColor) {
-                styles.push(`backgroundColor-${htmlElement.style.backgroundColor}`);
-            }
-            if (htmlElement.style.fontFamily) {
-                styles.push(`fontfamily-${htmlElement.style.fontFamily}`);
-            }
-            if (htmlElement.style.fontSize) {
-                styles.push(`fontsize-${htmlElement.style.fontSize}`);
-            }
-            if (htmlElement.tagName === 'SUB') {
-                styles.push('SUBSCRIPT');
-            }
-            if (htmlElement.tagName === 'SUP') {
-                styles.push('SUPERSCRIPT');
-            }
-            return styles.length ? Style(styles) : null;
-        },
-        
-    };
-
-    const contentState = stateFromHTML(content, importOptions);
-    
-    return EditorState.createWithContent(contentState);
-};
-
-const getFormat = (editorStateData: EditorState) => {
-    const style = editorStateData.getCurrentInlineStyle();
-    const format: IDraftElementFormats = {
-        bold: style.has('BOLD'),
-        italic: style.has('ITALIC'),
-        underline: style.has('UNDERLINE'),
-        subScript: style.has('SUBSCRIPT'),
-        superScript: style.has('SUPERSCRIPT'),
-    };
-    style.forEach((styleKey) => {
-        if (styleKey) {
-            if (styleKey.includes('color-')) {
-                const [, color] = styleKey.split('color-');
-                format.color = color;
-            } else if (styleKey.includes('backgroundColor-')) {
-                const [, backgroundColor] = styleKey.split('backgroundColor-');
-                format.backgroundColor = backgroundColor;
-            } else if (styleKey.includes('fontfamily-')) {
-                const [, font] = styleKey.split('fontfamily-');
-                try {
-                    format.font = JSON.parse(font);
-                } catch (error) {
-                    format.font = font;
-                }
-            } else if (styleKey.includes('fontsize-')) {
-                const [, size] = styleKey.split('fontsize-');
-                format.size = size;
-            }
-        }
-    });
-
-    return format;
-};
-
-const formatText = (editorState: EditorState, formatType: string, value: string) => {
-    const selection = editorState.getSelection();
-    let contentState = editorState.getCurrentContent();
-
-    const currentStyleBefore = editorState.getCurrentInlineStyle();
-
-    currentStyleBefore.forEach((color) => {
-        if (color && color.includes(`${formatType}-`)) {
-            contentState = Modifier.removeInlineStyle(contentState, selection, color);
-        }
-    });
-
-    let nextEditorState = EditorState.push(editorState, contentState, 'change-inline-style');
-
-    const currentStyle = editorState.getCurrentInlineStyle();
-    if (selection.isCollapsed()) {
-        nextEditorState = currentStyle.reduce(
-            (state, color) => RichUtils.toggleInlineStyle(state, color),
-            nextEditorState,
-        );
+const PopOverConatiner = (props) => {
+    debugger;
+    const boudingRect = props.store.getReferenceElement()?.getBoundingClientRect();
+    if (!boudingRect) {
+        return null
     }
-    if (!currentStyle.has(value)) {
-        nextEditorState = RichUtils.toggleInlineStyle(nextEditorState, value);
-    }
-    return nextEditorState;
-};
+    const style: React.CSSProperties = boudingRect ? { position: boudingRect ? "fixed" : null, left: boudingRect?.left + 20, top: boudingRect?.top + 10, backgroundColor: "#fff", zIndex: 1000, boxShadow: '0px 4px 4px rgb(0 0 0 / 25%)', maxHeight: '160px', overflowY: "auto" } : null;
+    return (<div style={style}>{props.children}</div>)
+}
 
-const getContentFromEditorState = (editorStateUpdated: EditorState) => {
-    const rawContentState = convertToRaw(editorStateUpdated.getCurrentContent());
-    return draftToHtml(rawContentState);
-};
-
-const convertToHTMLString = (editorState) =>{
+const convertToHTMLString = (editorState) => {
     return convertToHTML({
         styleToHTML: (style) => {
-          if (style === 'BOLD') {
-            return <span style={{color: 'blue'}} />;
-          }
+            if (style === 'BOLD') {
+                return <span style={{ color: 'blue' }} />;
+            }
         },
         blockToHTML: (block) => {
-          if (block.type === 'PARAGRAPH') {
-            return <p />;
-          }
+            if (block.type === 'PARAGRAPH') {
+                return <p />;
+            }
         },
         entityToHTML: (entity, originalText) => {
-          if (entity.type === 'mention') {
-            return <span className="mention" style={{backgroundColor: 'blue'}}data-value={JSON.stringify({...entity.data.mention, image:"", avatar:""})} >{originalText}</span>;
-          } else if(entity.type === '#mention'){
-            return <span className="hash-mention" style={{backgroundColor: 'orange'}} data-value={JSON.stringify({...entity.data.mention, image:"", avatar:""})} >{originalText}</span>
-          }
-          return originalText;
+            if (entity.type === 'mention') {
+                return <a href={entity.data.mention.link} className="mention" style={{ backgroundColor: 'blue' }} data-value={JSON.stringify({ ...entity.data.mention, image: "", avatar: "" })} >{originalText}</a>;
+            } else if (entity.type === '#mention') {
+                return <a href={entity.data.mention.link} className="hash-mention" style={{ backgroundColor: 'orange' }} data-value={JSON.stringify({ ...entity.data.mention, image: "", avatar: "" })} >{originalText}</a>
+            }
+            return originalText;
         }
     })(editorState.getCurrentContent());
-}
-
-const convertFromHTMLString = (html) => {
-    if(!html) {
-        html = ""
-    }
-    return convertFromHTML({
-        htmlToStyle: (nodeName, node, currentStyle) => {
-            if (nodeName === 'span' && node.style.color === 'blue') {
-                return currentStyle.add('BLUE');
-            } else {
-                return currentStyle;
-            }
-        },
-        htmlToEntity: (nodeName, node, createEntity) => {
-            if (nodeName === 'span' && node.classList.contains('mention')) {
-                const data = JSON.parse(node.dataset.value);
-                return createEntity(
-                    'MENTION',
-                    'IMMUTABLE',
-                    {mention: {name:data.name, ...data}}
-                )
-            } else if(nodeName === 'span' && node.classList.contains('hash-mention')) {
-                const data = JSON.parse(node.dataset.value);
-                return createEntity(
-                    'MENTION',
-                    'IMMUTABLE',
-                    {mention: {name:data.name, ...data}}
-                )
-            }
-        },
-        htmlToBlock: (nodeName, node) => {
-            if (nodeName === 'blockquote') {
-                return {
-                    type: 'blockquote',
-                    data: {}
-                };
-            }
-        }
-    })(html);
 }
 class DraftEditor extends Component<IDraftEditorProps, any> {
     private mentionSuggestionList: any
     constructor(props: IDraftEditorProps) {
         super(props);
         const { initialContent, peopleSuggestion, showMention } = props;
-        this.mentionSuggestionList = null;   
+        this.mentionSuggestionList = null;
         this.state = {
             editorState: EditorState.createWithContent(convertFromHTMLString(initialContent)),
             currentFormat: null,
@@ -248,10 +98,10 @@ class DraftEditor extends Component<IDraftEditorProps, any> {
             peopleSearchOpen: false,
             suggestions: props.valueSuggestion,
         }
-        if(showMention && (showMention.people || showMention.value)) {
+        if (showMention && (showMention.people || showMention.value)) {
             this.MentionComponents();
         }
-    }    
+    }
 
 
     sendFormat = (nextEditorState: EditorState) => {
@@ -283,7 +133,7 @@ class DraftEditor extends Component<IDraftEditorProps, any> {
         this.updateData(editorStateUpdated);
     };
 
-    
+
     updateData = (editorStateUpdated: EditorState) => {
         const { onContentTextChange } = this.props;
         this.setState({
@@ -291,20 +141,22 @@ class DraftEditor extends Component<IDraftEditorProps, any> {
         });
         const rawData = convertToRaw(editorStateUpdated.getCurrentContent());
         const mentionList = []
-        Object.keys(rawData.entityMap).forEach(key=> {
-            if(rawData.entityMap[key].type === '#mention') {
+        Object.keys(rawData.entityMap).forEach(key => {
+            if (rawData.entityMap[key].type === '#mention') {
                 return;
             }
-            mentionList.push({"emailAddress" : rawData.entityMap[key].data.mention.value});
+            mentionList.push({ "emailAddress": rawData.entityMap[key].data.mention.value });
 
-        })        
+        })
         const value = rawData.blocks.map(block => (!block.text.trim() && '\n') || block.text).join('\n');
-        onContentTextChange?.({formattedText: value,
-        value: convertToHTMLString(editorStateUpdated), mentionList })
+        onContentTextChange?.({
+            formattedText: value,
+            value: convertToHTMLString(editorStateUpdated), mentionList
+        })
         this.sendFormat(editorStateUpdated);
     };
 
-    onEditorTextChange= (editorStateUpdated:EditorState) => {
+    onEditorTextChange = (editorStateUpdated: EditorState) => {
         const html = convertToHTMLString(editorStateUpdated)
         const editorState = EditorState.createWithContent(convertFromHTMLString(html))
 
@@ -325,21 +177,21 @@ class DraftEditor extends Component<IDraftEditorProps, any> {
 
     MentionComponents = () => {
         const { showMention } = this.props;
-        const mentionPlugin_PREFIX_ONE = showMention.people ? createMentionPlugin({ mentionTrigger: MENTION_SUGGESTION_NAME.PREFIX_ONE }): {MentionSuggestions: null};
-        const mentionPlugin_PREFIX_TWO = showMention.value ? createMentionPlugin({ mentionTrigger: MENTION_SUGGESTION_NAME.PREFIX_TWO }): {MentionSuggestions: null};
+        const mentionPlugin_PREFIX_ONE = showMention.people ? createMentionPlugin({ mentionTrigger: MENTION_SUGGESTION_NAME.PREFIX_ONE, supportWhitespace: false, entityMutability: 'IMMUTABLE' }) : { MentionSuggestions: null };
+        const mentionPlugin_PREFIX_TWO = showMention.value ? createMentionPlugin({ mentionTrigger: MENTION_SUGGESTION_NAME.PREFIX_TWO, supportWhitespace: false, entityMutability: 'IMMUTABLE' }) : { MentionSuggestions: null };
 
 
         // eslint-disable-next-line no-shadow
         const { MentionSuggestions } = mentionPlugin_PREFIX_ONE;
-        const { MentionSuggestions : ValueSuggestion } = mentionPlugin_PREFIX_TWO;
+        const { MentionSuggestions: ValueSuggestion } = mentionPlugin_PREFIX_TWO;
 
-        
+
         // eslint-disable-next-line no-shadow
         const plugins = [];
-        if(showMention.people) {
+        if (showMention.people) {
             plugins.push(mentionPlugin_PREFIX_ONE)
         }
-        if(showMention.value) {
+        if (showMention.value) {
             plugins.push(mentionPlugin_PREFIX_TWO)
         }
 
@@ -347,21 +199,20 @@ class DraftEditor extends Component<IDraftEditorProps, any> {
         return this.mentionSuggestionList;
     }
 
-    onOpenChange =(searchKey: string) =>  (_open: boolean) => {
-        const { onMentionInput } = this.props;
+    onOpenChange = (searchKey: string) => (_open: boolean) => {
         this.setState({ [searchKey]: _open })
     };
-    onSearchChange = ({trigger, value }: {trigger:string, value: string }) => {
+    onSearchChange = ({ trigger, value }: { trigger: string, value: string }) => {
         const { onMentionInput, valueSuggestion } = this.props;
-        if(trigger === MENTION_SUGGESTION_NAME.PREFIX_ONE){
+        if (trigger === MENTION_SUGGESTION_NAME.PREFIX_ONE) {
             onMentionInput(value);
-        }else{
-            this.setState({ suggestions: defaultSuggestionsFilter(value,  valueSuggestion)})
+        } else {
+            this.setState({ suggestions: defaultSuggestionsFilter(value, valueSuggestion) })
         }
     };
     render() {
-        const { textAlignment, showToolbar,  peopleSuggestion, isMentionLoading  } = this.props;
-        const { editorState, currentFormat, peopleSearchOpen, valueSearchOpen, suggestions } = this.state;        
+        const { textAlignment, showToolbar, peopleSuggestion, isMentionLoading } = this.props;
+        const { editorState, currentFormat, peopleSearchOpen, valueSearchOpen, suggestions } = this.state;
         const MentionComp = this.mentionSuggestionList?.MentionSuggestions
         const ValueMentionComp = this.mentionSuggestionList?.ValueSuggestion
 
@@ -369,6 +220,12 @@ class DraftEditor extends Component<IDraftEditorProps, any> {
         return (
             <Fragment>
                 {showToolbar && <DraftToolbar currentFormat={currentFormat} setFormat={this.setFormat} />}
+                {/* <div
+                    className={editorStyles.editor}
+                    onClick={() => {
+                        ref.current!.focus();
+                    }}
+                    > */}
                 <Editor
                     customStyleFn={resolveCustomStyleMap}
                     preserveSelectionOnBlur
@@ -380,42 +237,45 @@ class DraftEditor extends Component<IDraftEditorProps, any> {
                     customStyleMap={CUSTOM_STYLE_MAP}
                     plugins={this.mentionSuggestionList?.plugins}
                 />
-                
-                
-                    <div className="list_container">
-                        {peopleSearchOpen && isMentionLoading && (
-                            <ul >
-                                <div className="menu-loading"></div>
-                            </ul>
-                        )}
-                        {peopleSearchOpen && !isMentionLoading && peopleSuggestion.length === 0 &&  (
-                            <ul style={{padding:"0 10px"}}>
-                                No Data found
-                            </ul>
-                        )}
-                       { MentionComp && !isMentionLoading && (<MentionComp
-                            open={peopleSearchOpen}
-                            onOpenChange={this.onOpenChange('peopleSearchOpen')}
-                            suggestions={peopleSuggestion}
-                            onSearchChange={this.onSearchChange}                            
-                            entryComponent={SuggestionList}
-                            popoverContainer={({ children }) => <div>{children}</div>}
-                        />
-                       )}
-                        { ValueMentionComp && 
-                            (<ValueMentionComp
-                                open={valueSearchOpen}
-                                onOpenChange={this.onOpenChange('valueSearchOpen')}
-                                suggestions={suggestions}
-                                onSearchChange={this.onSearchChange}
-                                onAddMention={() => {
-                                    // get the mention object selected
-                                }}
-                                entryComponent={SuggestionList}
-                                popoverContainer={({ children }) => <div>{children}</div>}
-                            />)
-                        }
-                    </div>            
+                <div className="list_container">
+                    {peopleSearchOpen && isMentionLoading && (
+                        <ul >
+                            <div className="menu-loading"></div>
+                        </ul>
+                    )}
+                    {peopleSearchOpen && !isMentionLoading && peopleSuggestion.length === 0 && (
+                        <ul style={{ padding: "0 10px" }}>
+                            No Data found
+                        </ul>
+                    )}
+                </div>
+                {MentionComp && !isMentionLoading && (<MentionComp
+                    open={peopleSearchOpen}
+                    onOpenChange={this.onOpenChange('peopleSearchOpen')}
+                    suggestions={peopleSuggestion}
+                    onSearchChange={this.onSearchChange}
+                    entryComponent={SuggestionList}
+                    // onAddMention={() => {
+                    //     this.onOpenChange('peopleSearchOpen')(false)
+                    // }}
+                    popoverContainer={PopOverConatiner}
+                />
+                )}
+                {ValueMentionComp &&
+                    (<ValueMentionComp
+                        open={valueSearchOpen}
+                        onOpenChange={this.onOpenChange('valueSearchOpen')}
+                        suggestions={suggestions}
+                        onSearchChange={this.onSearchChange}
+                        // onAddMention={() => {
+                        //     this.onOpenChange('valueSearchOpen')(false)
+                        // }}
+                        entryComponent={SuggestionList}
+                        popoverContainer={PopOverConatiner}
+                    />)
+                }
+                {/* </div> */}
+                {/* </>             */}
             </Fragment>
         );
     }
