+++ src/Styles/Mention.less
@@ -29,29 +29,28 @@
         width: 100%;
     }
 
-    .list_item{
-        display: flex;
-        align-items: center;
-        padding: 5px 10px;
-        &:hover{
-            background-color: #e6f3ff;
-        }
-
-    }
-    .list_icon{
-        display: flex;
-    }
-    .list-title{
-        font-size: 12px;
-        color: #666;
-    }
-    img{
-        width: 30px;
-        height: 30px;
-        border-radius: 50%;
-        margin-right: 10px;
+}
+.list_item{
+    display: flex;
+    align-items: center;
+    padding: 5px 10px;
+    &:hover{
+        background-color: #e6f3ff;
     }
 
+}
+.list_icon{
+    display: flex;
+}
+.list-title{
+    font-size: 12px;
+    color: #666;
+}
+img{
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+    margin-right: 10px;
 }
 
 .editor {
