+++ src/Styles/index.ts
@@ -2,6 +2,10 @@ import '../css/ToolbarButton.css';
 import '../css/DraftEditor.css';
 import '../css/Icon.css';
 import '../css/Mention.css';
+const  mentionsStyles = require('../css/MentionStyle.css');
+
+export { mentionsStyles }
+// import '../css/Mention.css';
 
 //To Run Local 
 // import './ToolbarButton.less';
